requirements.txt
aiogram
asyncpg
sqlalchemy[asyncio]
alembic
python-dotenv
pytest
pytz
Babel
aiocron
